requirements.txt
pandas>=2.0.0
pandas-datareader>=0.10.0
requests>=2.31.0
matplotlib>=3.7.0
seaborn>=0.12.0
jupyter>=1.0.0
pathlib2>=2.3.0
streamlit>=1.28.0
plotly>=5.15.0
numpy>=1.24.0
python-dateutil>=2.8.0
lru-dict>=1.1.8
psutil>=5.9.0
memory-profiler>=0.61.0
aiohttp>=3.8.0
asyncio-throttle>=1.0.2
beautifulsoup4>=4.12.0
textblob>=0.17.1
nltk>=3.8.1
websockets>=11.0.0
redis>=4.5.0

xgboost>=2.0.0
scikit-learn>=1.3.0
joblib>=1.3.0
cvxpy>=1.3.0
scipy>=1.10.0

# セキュリティ関連
PyJWT>=2.8.0
cryptography>=41.0.0
bcrypt>=4.0.0

# テスト関連
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0

# モニタリング・ログ関連
structlog>=23.1.0
prometheus-client>=0.17.0

# データベース関連
sqlalchemy>=2.0.0
alembic>=1.11.0

# Streamlit Cloud用の追加依存関係
watchdog>=3.0.0 